[SYM-19] Cost + time caps per binding (#121)
* Enforce acceptance cost and time caps

* Close acceptance runner stream after terminal event

+++ src/symphony/agent/runner.py
@@ -27,7 +27,7 @@ class RunnerSpec:
     )
     command: list[str]  # ["claude", "--print", "--output-format", "stream-json", ...]
     env: dict[str, str] = field(default_factory=dict)
-    stall_secs: int = 300
+    stall_secs: float = 300
     stage: str = ""  # implement|review|merge — telemetry only
